#482 Create auth.utils.ts
